Move config_manage out of webapps/ into config/
First pass.
