+++ Makefile
@@ -10,7 +10,7 @@ RELEASE_NEXT:=16.04
 RELEASE_NEXT_BRANCH:=dev
 RELEASE_UPSTREAM:=upstream
 MY_UPSTREAM:=origin
-CONFIG_MANAGE=$(IN_VENV) python lib/galaxy/webapps/config_manage.py
+CONFIG_MANAGE=$(IN_VENV) python lib/galaxy/config/config_manage.py
 PROJECT_URL?=https://github.com/galaxyproject/galaxy
 DOCS_DIR=doc
 DOC_SOURCE_DIR=$(DOCS_DIR)/source
